test: update test imports to match error subpackage reorganization
Redirect test file imports from deleted flat error files to the new
errors.base and errors.combined subpackage paths.
