+++ tests/forging_blocks/application/ports/inbound/test_validation_port.py
@@ -1,8 +1,8 @@
 import pytest
 
 from forging_blocks.application.ports.inbound.validation_port import ValidationPort
+from forging_blocks.foundation.errors import RuleViolated, RuleViolationError
 from forging_blocks.foundation.errors.core import ErrorMessage
-from forging_blocks.foundation.errors.rule_violation_error import RuleViolationError
 
 
 @pytest.mark.unit
@@ -10,7 +10,7 @@ class TestValidationPort:
     async def test_when_concrete_implementation_then_returns_command_errors(self) -> None:
         class StrictValidator(ValidationPort):
             async def validate_command(self, command: object) -> list[RuleViolationError]:
-                return [RuleViolationError(ErrorMessage("invalid"))]
+                return [RuleViolated(ErrorMessage("invalid"))]
 
             async def validate_query(self, query: object) -> list[RuleViolationError]:
                 del query
@@ -20,7 +20,7 @@ class TestValidationPort:
         errors = await service.validate_command("data")
 
         assert len(errors) == 1
-        assert str(errors[0]) == "RuleViolationError: invalid"
+        assert str(errors[0]) == "RuleViolated: invalid"
 
     async def test_when_valid_query_then_returns_empty(self) -> None:
         class PermissiveValidator(ValidationPort):
