+++ tests/forging_blocks/domain/errors/test_draft_entity_is_not_hashable_error.py
@@ -3,7 +3,7 @@ import pytest
 from forging_blocks.domain import (
     DraftEntityIsNotHashableError,
 )
-from forging_blocks.foundation.errors.error import Error
+from forging_blocks.foundation.errors.base.error import Error
 
 
 @pytest.mark.unit
